No need to write a int, using a byte instead.

+++ SiFEUPMobile/src/pt/up/beta/mobile/utils/ParcelUtils.java
@@ -7,13 +7,13 @@ public class ParcelUtils {
 	}
 
 	public static void writeString(Parcel dest, String value) {
-		dest.writeInt(value != null ? 1 : 0);
+		dest.writeByte((byte) (value != null ? 1 : 0));
 		if (value != null)
 			dest.writeString(value);
 	}
 
 	public static String readString(Parcel in) {
-		if (in.readInt() == 1)
+		if (in.readByte() == 1)
 			return in.readString();
 		return null;
 	}
